requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
openai>=1.0.0
easyocr>=1.7.0
Pillow>=10.0.0
PyMuPDF>=1.23.0 